Check for empty table before trying to clear it

+++ SQLServerDB/interview_question_Table.cs
@@ -217,6 +217,9 @@ namespace SQLServerDB
         /// </summary>
         public void Clear_Database_Table()
         {
+            if (CountRows() == 0)
+                return;
+
             string strCommand = "DELETE FROM " + theTable;
             if (!DBUtils.ExecuteSqlNonQuery(strCommand))
                 LogManager.writeToLog("ExecuteSqlNonQuery returned: FALSE; in interview_question_Table.cs:Clear_Database_Table. ExecuteSqlNonQuery()");
